Convert examples to ts
