+++ examples/run-all-examples.sh
@@ -1,13 +1,10 @@
 #/bin/bash
-set -e
+set -ex
 
-node ../cli.js --json gl.json5
-node ../cli.js --json image.json5
-node ../cli.js --json losslesscut.json5
-node ../cli.js --json resizeHorizontal.json5
-node ../cli.js --json resizeVertical.json5
-node ../cli.js --json speedTest.json5
-node ../cli.js --json subtitle.json5
-node ../cli.js --json transitionEasing.json5
-node ../cli.js --json transparentGradient.json5
-node ../cli.js --json commonFeatures.json5
+node ../dist/cli.js --json gl.json5
+node ../dist/cli.js --json image.json5
+node ../dist/cli.js --json speedTest.json5
+node ../dist/cli.js --json subtitle.json5
+node ../dist/cli.js --json transitionEasing.json5
+node ../dist/cli.js --json transparentGradient.json5
+node ../dist/cli.js --json commonFeatures.json5
